feat: Add support for AWS Bedrock in Assistant
- Implement AWS Bedrock adapter for Assistant
- Create new adapters for various LLMs (Anthropic, Google Gemini, MistralAI, Ollama, OpenAI)
- Refactor Assistant to use LLM-specific adapters
- Add base adapter class for common functionality
- Introduce AWS Bedrock message class

This change enhances the Assistant class to support AWS Bedrock and improves
the overall architecture by introducing LLM-specific adapters for better
extensibility and maintainability.
