+++ lib/langchain/assistants/assistant.rb
@@ -1,5 +1,7 @@
 # frozen_string_literal: true
 
+require_relative "llm/adapter"
+
 module Langchain
   # Assistants are Agent-like objects that leverage helpful instructions, LLMs, tools and knowledge to respond to user queries.
   # Assistants can be configured with an LLM of your choice, any vector search database and easily extended with additional tools.
@@ -411,323 +413,6 @@ module Langchain
 
     def available_tool_names
       llm_adapter.available_tool_names(tools)
-    end
-
-    # TODO: Fix the message truncation when context window is exceeded
-
-    module LLM
-      class Adapter
-        def self.build(llm)
-          case llm
-          when Langchain::LLM::Anthropic
-            Adapters::Anthropic.new
-          when Langchain::LLM::GoogleGemini, Langchain::LLM::GoogleVertexAI
-            Adapters::GoogleGemini.new
-          when Langchain::LLM::MistralAI
-            Adapters::MistralAI.new
-          when Langchain::LLM::Ollama
-            Adapters::Ollama.new
-          when Langchain::LLM::OpenAI
-            Adapters::OpenAI.new
-          else
-            raise ArgumentError, "Unsupported LLM type: #{llm.class}"
-          end
-        end
-      end
-
-      module Adapters
-        class Base
-          def build_chat_params(tools:, instructions:, messages:, tool_choice:)
-            raise NotImplementedError, "Subclasses must implement build_chat_params"
-          end
-
-          def extract_tool_call_args(tool_call:)
-            raise NotImplementedError, "Subclasses must implement extract_tool_call_args"
-          end
-
-          def build_message(role:, content: nil, image_url: nil, tool_calls: [], tool_call_id: nil)
-            raise NotImplementedError, "Subclasses must implement build_message"
-          end
-        end
-
-        class Ollama < Base
-          def build_chat_params(tools:, instructions:, messages:, tool_choice:)
-            params = {messages: messages}
-            if tools.any?
-              params[:tools] = build_tools(tools)
-            end
-            params
-          end
-
-          def build_message(role:, content: nil, image_url: nil, tool_calls: [], tool_call_id: nil)
-            warn "Image URL is not supported by Ollama currently" if image_url
-
-            Langchain::Messages::OllamaMessage.new(role: role, content: content, tool_calls: tool_calls, tool_call_id: tool_call_id)
-          end
-
-          # Extract the tool call information from the OpenAI tool call hash
-          #
-          # @param tool_call [Hash] The tool call hash
-          # @return [Array] The tool call information
-          def extract_tool_call_args(tool_call:)
-            tool_call_id = tool_call.dig("id")
-
-            function_name = tool_call.dig("function", "name")
-            tool_name, method_name = function_name.split("__")
-
-            tool_arguments = tool_call.dig("function", "arguments")
-            tool_arguments = if tool_arguments.is_a?(Hash)
-              Langchain::Utils::HashTransformer.symbolize_keys(tool_arguments)
-            else
-              JSON.parse(tool_arguments, symbolize_names: true)
-            end
-
-            [tool_call_id, tool_name, method_name, tool_arguments]
-          end
-
-          def available_tool_names(tools)
-            build_tools(tools).map { |tool| tool.dig(:function, :name) }
-          end
-
-          def allowed_tool_choices
-            ["auto", "none"]
-          end
-
-          private
-
-          def build_tools(tools)
-            tools.map { |tool| tool.class.function_schemas.to_openai_format }.flatten
-          end
-        end
-
-        class OpenAI < Base
-          def build_chat_params(tools:, instructions:, messages:, tool_choice:)
-            params = {messages: messages}
-            if tools.any?
-              params[:tools] = build_tools(tools)
-              params[:tool_choice] = build_tool_choice(tool_choice)
-            end
-            params
-          end
-
-          def build_message(role:, content: nil, image_url: nil, tool_calls: [], tool_call_id: nil)
-            Langchain::Messages::OpenAIMessage.new(role: role, content: content, image_url: image_url, tool_calls: tool_calls, tool_call_id: tool_call_id)
-          end
-
-          # Extract the tool call information from the OpenAI tool call hash
-          #
-          # @param tool_call [Hash] The tool call hash
-          # @return [Array] The tool call information
-          def extract_tool_call_args(tool_call:)
-            tool_call_id = tool_call.dig("id")
-
-            function_name = tool_call.dig("function", "name")
-            tool_name, method_name = function_name.split("__")
-
-            tool_arguments = tool_call.dig("function", "arguments")
-            tool_arguments = if tool_arguments.is_a?(Hash)
-              Langchain::Utils::HashTransformer.symbolize_keys(tool_arguments)
-            else
-              JSON.parse(tool_arguments, symbolize_names: true)
-            end
-
-            [tool_call_id, tool_name, method_name, tool_arguments]
-          end
-
-          def build_tools(tools)
-            tools.map { |tool| tool.class.function_schemas.to_openai_format }.flatten
-          end
-
-          def allowed_tool_choices
-            ["auto", "none"]
-          end
-
-          def available_tool_names(tools)
-            build_tools(tools).map { |tool| tool.dig(:function, :name) }
-          end
-
-          private
-
-          def build_tool_choice(choice)
-            case choice
-            when "auto"
-              choice
-            else
-              {"type" => "function", "function" => {"name" => choice}}
-            end
-          end
-        end
-
-        class MistralAI < Base
-          def build_chat_params(tools:, instructions:, messages:, tool_choice:)
-            params = {messages: messages}
-            if tools.any?
-              params[:tools] = build_tools(tools)
-              params[:tool_choice] = build_tool_choice(tool_choice)
-            end
-            params
-          end
-
-          def build_message(role:, content: nil, image_url: nil, tool_calls: [], tool_call_id: nil)
-            warn "Image URL is not supported by MistralAI currently" if image_url
-
-            Langchain::Messages::MistralAIMessage.new(role: role, content: content, tool_calls: tool_calls, tool_call_id: tool_call_id)
-          end
-
-          # Extract the tool call information from the OpenAI tool call hash
-          #
-          # @param tool_call [Hash] The tool call hash
-          # @return [Array] The tool call information
-          def extract_tool_call_args(tool_call:)
-            tool_call_id = tool_call.dig("id")
-
-            function_name = tool_call.dig("function", "name")
-            tool_name, method_name = function_name.split("__")
-
-            tool_arguments = tool_call.dig("function", "arguments")
-            tool_arguments = if tool_arguments.is_a?(Hash)
-              Langchain::Utils::HashTransformer.symbolize_keys(tool_arguments)
-            else
-              JSON.parse(tool_arguments, symbolize_names: true)
-            end
-
-            [tool_call_id, tool_name, method_name, tool_arguments]
-          end
-
-          def build_tools(tools)
-            tools.map { |tool| tool.class.function_schemas.to_openai_format }.flatten
-          end
-
-          def allowed_tool_choices
-            ["auto", "none"]
-          end
-
-          def available_tool_names(tools)
-            build_tools(tools).map { |tool| tool.dig(:function, :name) }
-          end
-
-          private
-
-          def build_tool_choice(choice)
-            case choice
-            when "auto"
-              choice
-            else
-              {"type" => "function", "function" => {"name" => choice}}
-            end
-          end
-        end
-
-        class GoogleGemini < Base
-          def build_chat_params(tools:, instructions:, messages:, tool_choice:)
-            params = {messages: messages}
-            if tools.any?
-              params[:tools] = build_tools(tools)
-              params[:system] = instructions if instructions
-              params[:tool_choice] = build_tool_config(tool_choice)
-            end
-            params
-          end
-
-          def build_message(role:, content: nil, image_url: nil, tool_calls: [], tool_call_id: nil)
-            warn "Image URL is not supported by Google Gemini" if image_url
-
-            Langchain::Messages::GoogleGeminiMessage.new(role: role, content: content, tool_calls: tool_calls, tool_call_id: tool_call_id)
-          end
-
-          # Extract the tool call information from the Google Gemini tool call hash
-          #
-          # @param tool_call [Hash] The tool call hash, format: {"functionCall"=>{"name"=>"weather__execute", "args"=>{"input"=>"NYC"}}}
-          # @return [Array] The tool call information
-          def extract_tool_call_args(tool_call:)
-            tool_call_id = tool_call.dig("functionCall", "name")
-            function_name = tool_call.dig("functionCall", "name")
-            tool_name, method_name = function_name.split("__")
-            tool_arguments = tool_call.dig("functionCall", "args").transform_keys(&:to_sym)
-            [tool_call_id, tool_name, method_name, tool_arguments]
-          end
-
-          def build_tools(tools)
-            tools.map { |tool| tool.class.function_schemas.to_google_gemini_format }.flatten
-          end
-
-          def allowed_tool_choices
-            ["auto", "none"]
-          end
-
-          def available_tool_names(tools)
-            build_tools(tools).map { |tool| tool.dig(:name) }
-          end
-
-          private
-
-          def build_tool_config(choice)
-            case choice
-            when "auto"
-              {function_calling_config: {mode: "auto"}}
-            when "none"
-              {function_calling_config: {mode: "none"}}
-            else
-              {function_calling_config: {mode: "any", allowed_function_names: [choice]}}
-            end
-          end
-        end
-
-        class Anthropic < Base
-          def build_chat_params(tools:, instructions:, messages:, tool_choice:)
-            params = {messages: messages}
-            if tools.any?
-              params[:tools] = build_tools(tools)
-              params[:tool_choice] = build_tool_choice(tool_choice)
-            end
-            params[:system] = instructions if instructions
-            params
-          end
-
-          def build_message(role:, content: nil, image_url: nil, tool_calls: [], tool_call_id: nil)
-            warn "Image URL is not supported by Anthropic currently" if image_url
-
-            Langchain::Messages::AnthropicMessage.new(role: role, content: content, tool_calls: tool_calls, tool_call_id: tool_call_id)
-          end
-
-          # Extract the tool call information from the Anthropic tool call hash
-          #
-          # @param tool_call [Hash] The tool call hash, format: {"type"=>"tool_use", "id"=>"toolu_01TjusbFApEbwKPRWTRwzadR", "name"=>"news_retriever__get_top_headlines", "input"=>{"country"=>"us", "page_size"=>10}}], "stop_reason"=>"tool_use"}
-          # @return [Array] The tool call information
-          def extract_tool_call_args(tool_call:)
-            tool_call_id = tool_call.dig("id")
-            function_name = tool_call.dig("name")
-            tool_name, method_name = function_name.split("__")
-            tool_arguments = tool_call.dig("input").transform_keys(&:to_sym)
-            [tool_call_id, tool_name, method_name, tool_arguments]
-          end
-
-          def build_tools(tools)
-            tools.map { |tool| tool.class.function_schemas.to_anthropic_format }.flatten
-          end
-
-          def allowed_tool_choices
-            ["auto", "any"]
-          end
-
-          def available_tool_names(tools)
-            build_tools(tools).map { |tool| tool.dig(:name) }
-          end
-
-          private
-
-          def build_tool_choice(choice)
-            case choice
-            when "auto"
-              {type: "auto"}
-            when "any"
-              {type: "any"}
-            else
-              {type: "tool", name: choice}
-            end
-          end
-        end
-      end
     end
   end
 end
